+++ backend/map/__init__.py
@@ -1,0 +1,1 @@
+# This file can remain empty
